backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers>=3.10,<4.0
celery==5.3.4
redis==4.6.0
pytest>=6.2,<7.0
pytest-django>=4.5,<5.0
django-celery-beat==2.5.0
python-dotenv==1.0.0